+++ wemake_python_styleguide/constants.py
@@ -81,6 +81,8 @@ VARIABLE_NAMES_BLACKLIST: Final = frozenset((
     'objs',
     'some',
     'do',
+    'params',
+    'parameters',
 
     # Confuseables:
     'no',