+++ CHANGELOG.md
@@ -4,7 +4,7 @@ We follow Semantic Versions since the `0.1.0` release.
 We used to have incremental versioning before `0.1.0`.
 
 
-## WIP
+## 0.7.0
 
 ### Features
 
@@ -34,6 +34,7 @@ We used to have incremental versioning before `0.1.0`.
 - Forbids to have useless `continue` nodes inside the loops
 - Forbids to have useless nodes
 - Forbids to have useless `raise` statements
+- Adds `params` and `parameters` to black-listed names
 
 
 ### Bugfixes